requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
google-cloud-bigquery>=3.14.0
python-dateutil>=2.8.0
